requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Cache & Queue
redis>=5.0.0
celery>=5.3.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# ML
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
xgboost>=2.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.1.0
joblib>=1.3.0

# RAG
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
chromadb>=0.4.0
sentence-transformers>=2.2.0
rank-bm25>=0.2.2

# Qdrant + Hybrid Search
qdrant-client>=1.7.0
fastembed>=0.2.0

# Image Processing
opencv-python>=4.9.0
Pillow>=10.0.0

# Document Processing
pypdf>=3.17.0
beautifulsoup4>=4.12.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
ruff>=0.1.0
black>=23.12.0
mypy>=1.8.0
pre-commit>=3.6.0
